# [exporter]prometheusremotewrite] Fix data race by introducing pool of batch state