--- a/exporter/prometheusremotewriteexporter/exporter_test.go
+++ b/exporter/prometheusremotewriteexporter/exporter_test.go
@@ -1289,7 +1289,7 @@ func BenchmarkPushMetricsVaryingMetrics(b *testing.B) {
 }
 
 // benchmarkPushMetrics benchmarks the PushMetrics method with a given number of metrics.
-// If numMetrics is -1, it will benchmark with varying number of metrics, from 10 up to 10000.
+// If numMetrics is -1, it will benchmark with varying number of metrics, from 100 up to 1000000.
 func benchmarkPushMetrics(b *testing.B, numMetrics, numConsumers int) {
 	mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusOK)
@@ -1327,9 +1327,12 @@ func benchmarkPushMetrics(b *testing.B, numMetrics, numConsumers int) {
 	for n := 0; n < b.N; n++ {
 		actualNumMetrics := numMetrics
 		if numMetrics == -1 {
-			actualNumMetrics = int(math.Pow(10, float64(n%4+1)))
+			actualNumMetrics = int(math.Pow(100, float64(n%3+1)))
 		}
 		m := testdata.GenerateMetricsManyMetricsSameResource(actualNumMetrics)
+		for i := 0; i < m.MetricCount(); i++ {
+			m.ResourceMetrics().At(0).ScopeMetrics().At(0).Metrics().At(i).Sum().DataPoints().AppendEmpty().SetIntValue(int64(i))
+		}
 		metrics = append(metrics, m)
 	}