@Aneurysm9 2024-12-01
> This is strange to me. I don't see any degradations in memory allocations by removing the state, so I'm confused if the optimization ever made any difference or if I'm doing something wrong in the benchmark.

The [original issue](https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/34269) talked about batch sizes of 100-200k data points, so maybe 10k isn't enough to illustrate the issue?