@ArthurSens 2024-12-02
I've increased the amount of data used in the benchmarks, still no difference.

```console

arthursens$ benchstat main.txt withoutState.txt syncpool.txt 
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter
cpu: Apple M2 Pro
                            │  main.txt   │       withoutState.txt       │         syncpool.txt         │
                            │   sec/op    │   sec/op     vs base         │   sec/op     vs base         │
PushMetricsVaryingMetrics-2   676.0m ± 5%   685.0m ± 8%  ~ (p=0.240 n=6)   682.9m ± 6%  ~ (p=0.310 n=6)

                            │   main.txt   │       withoutState.txt        │         syncpool.txt          │
                            │     B/op     │     B/op      vs base         │     B/op      vs base         │
PushMetricsVaryingMetrics-2   431.0Mi ± 0%   431.2Mi ± 0%  ~ (p=0.937 n=6)   430.9Mi ± 0%  ~ (p=0.132 n=6)

                            │  main.txt   │        withoutState.txt        │           syncpool.txt            │
                            │  allocs/op  │  allocs/op   vs base           │  allocs/op   vs base              │
PushMetricsVaryingMetrics-2   4.334M ± 0%   4.334M ± 0%  ~ (p=1.000 n=6) ¹   4.334M ± 0%  +0.00% (p=0.002 n=6)
¹ all samples are equal

```

<details><summary>Individual benchmark runs</summary>

Main:

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter
cpu: Apple M2 Pro
BenchmarkPushMetricsVaryingMetrics-2   	     100	 641808841 ns/op	451837730 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 680346175 ns/op	452476772 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 673599777 ns/op	451837744 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 674459682 ns/op	451847461 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 677447471 ns/op	452476337 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 687872721 ns/op	452075896 B/op	 4333723 allocs/op
PASS
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter	543.110s

```

Alternative 1 (https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36600):

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter
cpu: Apple M2 Pro
BenchmarkPushMetricsVaryingMetrics-2   	     100	 632321270 ns/op	452227454 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 688034830 ns/op	451837198 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 685106043 ns/op	452075646 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 680661115 ns/op	451837636 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 684886080 ns/op	452237544 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 686308490 ns/op	452876716 B/op	 4333723 allocs/op
PASS
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter	544.852s
```

This branch:

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter
cpu: Apple M2 Pro
BenchmarkPushMetricsVaryingMetrics-2   	     100	 641808841 ns/op	451837730 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 680346175 ns/op	452476772 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 673599777 ns/op	451837744 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 674459682 ns/op	451847461 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 677447471 ns/op	452476337 B/op	 4333723 allocs/op
BenchmarkPushMetricsVaryingMetrics-2   	     100	 687872721 ns/op	452075896 B/op	 4333723 allocs/op
PASS
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter	543.110s

```
</details>