@dashpole 2024-12-02
Seems like a benchmarking problem.  Lets try to work with the original author in the other PR before we choose a path forward.