@ArthurSens 2024-12-05
> Seems like a benchmarking problem. 

That's for sure! I'm struggling to understand why the results look that way :/