@ArthurSens 2024-12-13
Okay, I found the problem. I'm creating a large request by creating an OTel metric with millions of Metrics with similar names and attributes within the same ResourceMetrics. This ends up in a very small `tsMap` [here](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/d72bbd21c83f4a65269b5e973f86459a23c94434/exporter/prometheusremotewriteexporter/exporter.go#L195). Even in the test case with 1 million datapoints we only get three entries in `tsMap`.

If we go a bit further and look at the code of `batchTimeSeries`, even if one single entry of tsMap holds 1 million datapoints (around 18MB) and the max batchSize is super small. Those 1 million datapoints become one single PRW request:

https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/d72bbd21c83f4a65269b5e973f86459a23c94434/exporter/prometheusremotewriteexporter/helper.go#L47-L54

That happens because `converTimeseriesToRequest` is not aware of maxBatchSize. I prefer that we fix this bug in a follow-up PR.

To progress the benchmark here, even with the bug, we can diversify the attributes in the original OTel metrics during the benchmark preparation, leading to millions of entries in tsMap. I can even decrease the amount of series again and it still works :)